Demodularization Script
