+++ index.html
@@ -87,7 +87,7 @@
                   <h4>JK Rowling - 1971</h4>
                 </div>
                 <div class="card__button">
-                  <span class="material-symbols-outlined"> done_outline </span>
+                  <span class="material-symbols-outlined"> undo </span>
                   <span class="material-symbols-outlined"> delete </span>
                 </div>
               </div>
@@ -106,7 +106,7 @@
                   <h4>JK Rowling - 1971</h4>
                 </div>
                 <div class="card__button">
-                  <span class="material-symbols-outlined"> undo </span>
+                  <span class="material-symbols-outlined"> done_outline </span>
                   <span class="material-symbols-outlined"> delete </span>
                 </div>
               </div>
